Up to date HR_Appraisal module

+++ custom_addons/hr_appraisal_objectives/models/hr_appraisal_goal_kpi.py
@@ -16,7 +16,7 @@ class HrAppraisalGoalKPI(models.Model):
     description = fields.Html("Description")
     target = fields.Html("Target/Goal")
     measurement_method = fields.Html("Measurement Method")
-    weight = fields.Float("Weight", default=1.0)
+    weight = fields.Integer("Weight", default=1)
 
     # --- Project integration ---
     project_id = fields.Many2one(
